🚚 move rocket common logic in presentation crate
