+++ api/Cargo.toml
@@ -70,6 +70,7 @@ domain = {path = "../domain", features = ["rocket"]}
 event-store = {path = "../event-store"}
 infrastructure = {path = "../infrastructure"}
 macros = {path = "../macros"}
+presentation = {path = "../presentation"}
 
 [dev-dependencies]
 async-std = {version = "1.12.0", features = ["attributes"]}
