--- api/src/presentation/http/config.rs
@@ -1,7 +1,0 @@
-use derive_getters::Getters;
-use serde::Deserialize;
-
-#[derive(Clone, Deserialize, Getters)]
-pub struct Config {
-	api_key: String,
-}
