+++ api/src/presentation/http/dto/mod.rs
@@ -1,0 +1,5 @@
+mod identity;
+pub use identity::IdentityInput;
+
+mod payout_settings;
+pub use payout_settings::PayoutSettingsInput;
