+++ api/src/presentation/http/roles.rs
@@ -1,0 +1,18 @@
+use crate::domain::{permissions, Permissions};
+use presentation::http::guards::Role;
+
+impl From<Role> for Box<dyn Permissions> {
+	fn from(role: Role) -> Self {
+		match role {
+			Role::Admin => permissions::of_admin(),
+			Role::IdentifiedUser {
+				lead_projects,
+				owned_budgets,
+			} => permissions::of_identified_user(
+				lead_projects.into_iter().map(Into::into).collect(),
+				owned_budgets.into_iter().map(Into::into).collect(),
+			),
+			Role::Public => permissions::of_anonymous(),
+		}
+	}
+}
