--- api/src/presentation/http/routes/dto/mod.rs
@@ -1,8 +1,0 @@
-mod role;
-pub use role::Role;
-
-pub mod api_key;
-pub mod identity;
-pub mod option_user_id;
-pub mod payout_settings;
-pub mod uuid;
