--- api/src/presentation/http/routes/dto/to_http_api_problem.rs
@@ -1,27 +1,0 @@
-use domain::*;
-use http_api_problem::{HttpApiProblem, StatusCode};
-
-pub(crate) trait ToHttpApiProblem {
-	fn to_http_api_problem(&self) -> HttpApiProblem;
-}
-
-impl ToHttpApiProblem for AggregateRootRepositoryError {
-	fn to_http_api_problem(&self) -> HttpApiProblem {
-		match self {
-			AggregateRootRepositoryError::NotFound =>
-				HttpApiProblem::new(StatusCode::NOT_FOUND).title(self.to_string()),
-			AggregateRootRepositoryError::EventStoreError(e) =>
-				HttpApiProblem::new(StatusCode::INTERNAL_SERVER_ERROR)
-					.title(self.to_string())
-					.detail(e.to_string()),
-		}
-	}
-}
-
-impl ToHttpApiProblem for EventStoreError {
-	fn to_http_api_problem(&self) -> HttpApiProblem {
-		HttpApiProblem::new(StatusCode::INTERNAL_SERVER_ERROR)
-			.title("Internal error")
-			.detail(self.to_string())
-	}
-}
