--- api/src/presentation/http/routes/dto/uuid.rs
@@ -1,16 +1,0 @@
-use derive_more::From;
-use rocket::request::FromParam;
-use serde::Deserialize;
-use std::str::FromStr;
-use uuid::Uuid;
-
-#[derive(Debug, Clone, Copy, Deserialize, From)]
-pub struct UuidParam(Uuid);
-
-impl<'a> FromParam<'a> for UuidParam {
-	type Error = uuid::Error;
-
-	fn from_param(param: &'a str) -> Result<Self, Self::Error> {
-		Uuid::from_str(param).map(UuidParam)
-	}
-}
