+++ api/src/presentation/http/routes/mod.rs
@@ -1,6 +1,1 @@
-pub mod cors;
-mod dto;
 pub mod graphql;
-pub mod health;
-
-pub use dto::*;
