+++ github-proxy/Cargo.toml
@@ -27,5 +27,6 @@ dotenv = "0.15.0"
 
 # Local dependecies
 infrastructure = {path = "../infrastructure"}
+presentation = {path = "../presentation"}
 
 [dev-dependencies]
