+++ github-proxy/src/main.rs
@@ -14,7 +14,7 @@ mod domain;
 mod infrastructure;
 mod presentation;
 
-use presentation::http;
+use self::presentation::http;
 use serde::Deserialize;
 
 #[derive(Deserialize)]
