+++ github-proxy/src/presentation/http/mod.rs
@@ -1,14 +1,13 @@
 use crate::domain::GithubService;
 use anyhow::Result;
-use infrastructure::github;
-use rocket::{Build, Rocket};
+use presentation::http;
 use std::sync::Arc;
 
-mod config;
 mod routes;
 
-pub async fn serve(github_client: Arc<github::Client>) -> Result<()> {
-	let _ = inject_app(rocket::custom(config::get()), github_client)
+pub async fn serve(github_service: Arc<dyn GithubService>) -> Result<()> {
+	let _ = rocket::custom(http::config::rocket("github-proxy/Rocket.toml"))
+		.manage(github_service)
 		.mount(
 			"/",
 			routes![
@@ -20,8 +19,4 @@ pub async fn serve(github_client: Arc<github::Client>) -> Result<()> {
 		.launch()
 		.await?;
 	Ok(())
-}
-
-fn inject_app(rocket: Rocket<Build>, github_service: Arc<dyn GithubService>) -> Rocket<Build> {
-	rocket.manage(github_service)
 }
