+++ presentation/Cargo.toml
@@ -6,3 +6,33 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+
+# Async
+async-trait = "0.1.56"
+
+# Core types
+uuid = {version = "0.8.2", default_features = false, features = [
+  "v4",
+  "serde",
+]}
+
+# Error
+thiserror = "1.0.31"
+
+# Http Server
+rocket = "0.5.0-rc.2"
+
+# Serde
+serde = {version = "1.0.137", features = ["derive"]}
+serde_json = {version = "1.0.81"}
+
+# Utils
+derive-getters = "0.2.0"
+
+# Local dependecies
+domain = {path = "../domain"}
+
+[dev-dependencies]
+async-std = {version = "1.12.0", features = ["attributes"]}
+rstest = "0.15.0"
+assert_matches = "1.5"
