+++ presentation/src/http/routes/cors.rs
@@ -1,0 +1,7 @@
+use rocket::http::Status;
+use std::path::PathBuf;
+
+#[options("/<_path..>")]
+pub fn options_preflight_handler(_path: PathBuf) -> Status {
+	Status::Ok
+}
