+++ presentation/src/lib.rs
@@ -1,14 +1,4 @@
-pub fn add(left: usize, right: usize) -> usize {
-	left + right
-}
+pub mod http;
 
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn it_works() {
-		let result = add(2, 2);
-		assert_eq!(result, 4);
-	}
-}
+#[macro_use]
+extern crate rocket;
